MT#59962 xmlrpc2di: deprecate int support
As long as the `TypeInt` is actually a long primitive,
and we tend to switch using the long type everywhere
within xml implementation, then
there is no need to support both `int` and `long`
types, just force users to be on `long` always.

Refactor the user code accordingly.

+++ apps/examples/tutorial/cc_acc_xmlrpc/CCAcc.cpp
@@ -90,7 +90,7 @@ int CCAcc::subtractCredit(string pin, int amount) {
    XmlRpcValue xmlArg;
    xmlArg[0][0]["methodName"] = "subtractCredit";
    xmlArg[0][0]["pin"] = pin;
-   xmlArg[0][0]["amount"] = amount;	
+   xmlArg[0][0]["amount"] = static_cast<long>(amount);
    DBG("subtractCredit pin# '%s', Seconds '%u'.\n", pin.c_str(),  
 amount );
    xmlrpccall.execute("subtractCredit", xmlArg, result);
